split lua main init
